refactor: 하나의 mutation.ts 파일로 통합
